Move TimeoutContext to services-client so it can be used in isomorphic packages
